% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{dabtab_viewer}
\alias{dabtab_viewer}
\title{Launch dabtab in the Rstudio viewer}
\usage{
dabtab_viewer(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch dabtab in the Rstudio viewer
}
\details{
See \url{https://dabtab-rstats.github.io/docs/} for documentation and tutorials
}
\examples{
\dontrun{
dabtab_viewer()
}
}
