% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabtab.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher and updater for Mac (.command)}
\usage{
mac_launcher(
  app = c("dabtab", "dabtab.basic"),
  port = 4444,
  pdir = Sys.getenv("HOME")
)
}
\arguments{
\item{app}{App to run when the desktop icon is double-clicked ("dabtab", "dabtab.basic", etc.). Default is "dabtab"}

\item{port}{Port to use for shiny::runApp (e.g, port = 4444)}

\item{pdir}{Project directory to use. Default is Sys.getenv("HOME")}
}
\description{
Create a launcher and updater for Mac (.command)
}
\details{
On Mac a file named 'dabtab.command' and one named 'update_dabtab.command' will be put on the desktop. Double-click the file to launch the specified dabtab app or update dabtab to the latest version
}
\examples{
\dontrun{
dabtab::mac_launcher()
}

}
